change file structure of page login register assign page
